pkg/apis/conditioner/v1alpha1/podcondition_types_test.go
/*
Copyright 2019 Itay Shakury.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package v1alpha1

import (
	"testing"

	"github.com/onsi/gomega"
	"golang.org/x/net/context"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
	"k8s.io/apimachinery/pkg/types"
)

func TestStorageConditioner(t *testing.T) {
	key := types.NamespacedName{
		Name:      "testcondition",
		Namespace: "default",
	}
	created := &PodCondition{
		ObjectMeta: metav1.ObjectMeta{
			Name:      "testcondition",
			Namespace: "default",
		},
		Spec: PodConditionSpec{
			Interval: 1,
			LabelSelector: metav1.LabelSelector{
				MatchLabels: map[string]string{
					"testing": "123",
				},
			},
			PrometheusSource: &PrometheusSource{
				ServerURL: "http://localhost:9090",
				Rule:      "vector(1) > 0",
			},
		},
	}
	g := gomega.NewGomegaWithT(t)

	// Test Create
	fetched := &PodCondition{}
	g.Expect(c.Create(context.TODO(), created)).NotTo(gomega.HaveOccurred())

	g.Expect(c.Get(context.TODO(), key, fetched)).NotTo(gomega.HaveOccurred())
	g.Expect(fetched).To(gomega.Equal(created))

	// Test Updating the Labels
	updated := fetched.DeepCopy()
	updated.Labels = map[string]string{"hello": "world"}
	g.Expect(c.Update(context.TODO(), updated)).NotTo(gomega.HaveOccurred())
	g.Expect(c.Get(context.TODO(), key, fetched)).NotTo(gomega.HaveOccurred())
	g.Expect(fetched).To(gomega.Equal(updated))

	// Test Delete
	g.Expect(c.Delete(context.TODO(), fetched)).NotTo(gomega.HaveOccurred())
	g.Expect(c.Get(context.TODO(), key, fetched)).To(gomega.HaveOccurred())
}
